Struct ethers_solc::cache::CacheEntry

pub struct CacheEntry {
    pub last_modification_date: u64,
    pub content_hash: String,
    pub source_name: PathBuf,
    pub solc_config: SolcConfig,
    pub imports: BTreeSet<PathBuf>,
    pub version_requirement: Option<String>,
    pub artifacts: BTreeMap<String, BTreeMap<Version, PathBuf>>,
}

A CacheEntry in the cache file represents a solidity file

A solidity file can contain several contracts, for every contract a separate Artifact is emitted. so the CacheEntry tracks the artifacts by name. A file can be compiled with multiple solc versions generating version specific artifacts.

Fields
last_modification_date: u64

the last modification time of this file

content_hash: String

hash to identify whether the content of the file changed

source_name: PathBuf

identifier name see crate::utils::source_name()

solc_config: SolcConfig

what config was set when compiling this file

imports: BTreeSet<PathBuf>

fully resolved imports of the file

all paths start relative from the project’s root: src/importedFile.sol

version_requirement: Option<String>

The solidity version pragma

artifacts: BTreeMap<String, BTreeMap<Version, PathBuf>>

all artifacts produced for this file

In theory a file can be compiled by different solc versions: A(<=0.8.10) imports C(>0.4.0) and B(0.8.11) imports C(>0.4.0) file C would be compiled twice, with 0.8.10 and 0.8.11, producing two different artifacts.

This map tracks the artifacts by name -> (Version -> PathBuf). This mimics the default artifacts directory structure

Implementations

impl CacheEntry

pub fn last_modified(&self) -> Duration

Returns the last modified timestamp Duration

pub fn find_artifact_path( &self, contract_name: impl AsRef<str> ) -> Option<&PathBuf>

Returns the artifact path for the contract name

use ethers_solc::cache::CacheEntry;
entry.find_artifact_path("Greeter");

pub fn read_last_modification_date(file: impl AsRef<Path>) -> Result<u64>

Reads the last modification date from the file’s metadata

pub fn retain_versions<'a, I>(&mut self, versions: I)
where I: IntoIterator<Item = &'a Version>,

Retains only those artifacts that match the provided versions.

Removes an artifact entry if none of its versions is included in the versions set.

pub fn contains_version(&self, version: &Version) -> bool

Returns true if the artifacts set contains the given version

pub fn artifacts_versions(&self) -> impl Iterator<Item = (&Version, &PathBuf)>

Iterator that yields all artifact files and their version

pub fn find_artifact( &self, contract: &str, version: &Version ) -> Option<&PathBuf>

Returns the artifact file for the contract and version pair

pub fn artifacts_for_version<'a>( &'a self, version: &'a Version ) -> impl Iterator<Item = &'a PathBuf> + 'a

Iterator that yields all artifact files and their version

pub fn artifacts(&self) -> impl Iterator<Item = &PathBuf>

Iterator that yields all artifact files

pub fn artifacts_mut(&mut self) -> impl Iterator<Item = &mut PathBuf>

Mutable iterator over all artifact files

pub fn all_artifacts_exist(&self) -> bool

Checks if all artifact files exist

pub fn join_artifacts_files(&mut self, base: impl AsRef<Path>)

Sets the artifact’s paths to base adjoined to the artifact’s path.

pub fn strip_artifact_files_prefixes(&mut self, base: impl AsRef<Path>)

Removes base from the artifact’s path