Struct ethers_solc::output::ProjectCompileOutput

pub struct ProjectCompileOutput<T: ArtifactOutput = ConfigurableArtifacts> { /* private fields */ }

Contains a mixture of already compiled/cached artifacts and the input set of sources that still need to be compiled.

Implementations

impl<T: ArtifactOutput> ProjectCompileOutput<T>

pub fn slash_paths(&mut self)

Converts all \\ separators in all paths to /

pub fn with_slashed_paths(self) -> Self

Convenience function fo Self::slash_paths()

pub fn artifact_ids( &self ) -> impl Iterator<Item = (ArtifactId, &T::Artifact)> + '_

All artifacts together with their contract file name and name <file name>:<name>.

This returns a chained iterator of both cached and recompiled contract artifacts.

Borrowed version of Self::into_artifacts.

pub fn into_artifacts(self) -> impl Iterator<Item = (ArtifactId, T::Artifact)>

All artifacts together with their contract file name and name <file name>:<name>

This returns a chained iterator of both cached and recompiled contract artifacts

Example

use std::collections::btree_map::BTreeMap;
use ethers_solc::ConfigurableContractArtifact;
use ethers_solc::{ArtifactId, Project};

let project = Project::builder().build().unwrap();
let contracts: BTreeMap<ArtifactId, ConfigurableContractArtifact> = project.compile().unwrap().into_artifacts().collect();

pub fn artifacts(&self) -> impl Iterator<Item = (String, &T::Artifact)>

This returns a chained iterator of both cached and recompiled contract artifacts that yields the contract name and the corresponding artifact

Example

use std::collections::btree_map::BTreeMap;
use ethers_solc::ConfigurableContractArtifact;
use ethers_solc::Project;

let project = Project::builder().build().unwrap();
let artifacts: BTreeMap<String, &ConfigurableContractArtifact> = project.compile().unwrap().artifacts().collect();

pub fn versioned_artifacts( &self ) -> impl Iterator<Item = (String, (&T::Artifact, &Version))>

This returns a chained iterator of both cached and recompiled contract artifacts that yields the contract name and the corresponding artifact with its version

Example

use std::collections::btree_map::BTreeMap;
use semver::Version;
use ethers_solc::ConfigurableContractArtifact;
use ethers_solc::Project;

let project = Project::builder().build().unwrap();
let artifacts: BTreeMap<String, (&ConfigurableContractArtifact, &Version)> = project.compile().unwrap().versioned_artifacts().collect();

pub fn artifacts_with_files( &self ) -> impl Iterator<Item = (&String, &String, &T::Artifact)> + '_

All artifacts together with their contract file and name as tuple (file, contract name, artifact)

This returns a chained iterator of both cached and recompiled contract artifacts

Borrowed version of Self::into_artifacts_with_files.

NOTE the file will be returned as is, see also Self::with_stripped_file_prefixes().

pub fn into_artifacts_with_files( self ) -> impl Iterator<Item = (String, String, T::Artifact)>

All artifacts together with their contract file and name as tuple (file, contract name, artifact)

This returns a chained iterator of both cached and recompiled contract artifacts

Example

use std::collections::btree_map::BTreeMap;
use ethers_solc::{ConfigurableContractArtifact, Project};

let project = Project::builder().build().unwrap();
let contracts: Vec<(String, String, ConfigurableContractArtifact)> = project.compile().unwrap().into_artifacts_with_files().collect();

NOTE the file will be returned as is, see also Self::with_stripped_file_prefixes()

pub fn into_artifacts_with_sources( self ) -> (BTreeMap<ArtifactId, T::Artifact>, VersionedSourceFiles)

All artifacts together with their ID and the sources of the project.

Note: this only returns the SourceFiles for freshly compiled contracts because, if not included in the Artifact itself (see crate::ConfigurableContractArtifact::source_file()), is only available via the solc CompilerOutput

pub fn with_stripped_file_prefixes(self, base: impl AsRef<Path>) -> Self

Strips the given prefix from all artifact file paths to make them relative to the given base argument

Example

Make all artifact files relative to the project’s root directory

use ethers_solc::Project;

let project = Project::builder().build().unwrap();
let output = project.compile().unwrap().with_stripped_file_prefixes(project.root());

pub fn output(self) -> AggregatedCompilerOutput

Get the (merged) solc compiler output

use std::collections::btree_map::BTreeMap;
use ethers_solc::artifacts::contract::Contract;
use ethers_solc::Project;

let project = Project::builder().build().unwrap();
let contracts: BTreeMap<String, Contract> =
    project.compile().unwrap().output().contracts_into_iter().collect();

pub fn has_compiled_contracts(&self) -> bool

Returns whether this type has a compiler output.

pub fn is_unchanged(&self) -> bool

Returns whether this type does not contain compiled contracts.

pub fn has_compiler_errors(&self) -> bool

Returns whether any errors were emitted by the compiler.

pub fn has_compiler_warnings(&self) -> bool

Returns whether any warnings were emitted by the compiler.

pub fn succeeded(self) -> Self

Panics if any errors were emitted by the compiler.

pub fn assert_success(&self)

Panics if any errors were emitted by the compiler.

pub fn cached_artifacts(&self) -> &Artifacts<T::Artifact>

Returns the set of Artifacts that were cached and got reused during crate::Project::compile()

pub fn compiled_artifacts(&self) -> &Artifacts<T::Artifact>

Returns the set of Artifacts that were compiled with solc in crate::Project::compile()

pub fn compiled_contracts_by_compiler_version( &self ) -> BTreeMap<Version, Vec<(String, Contract)>>

Returns a BTreeMap that maps the compiler version used during crate::Project::compile() to a Vector of tuples containing the contract name and the Contract

pub fn find_contract<'a>( &self, info: impl Into<ContractInfoRef<'a>> ) -> Option<&T::Artifact>

Removes the contract with matching path and name using the <path>:<contractname> pattern where path is optional.

If the path segment is None, then the first matching Contract is returned, see Self::remove_first

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
use ethers_solc::info::ContractInfo;

let  output = project.compile().unwrap();
let info = ContractInfo::new("src/Greeter.sol:Greeter");
let contract = output.find_contract(&info).unwrap();

pub fn find( &self, path: impl AsRef<str>, contract: impl AsRef<str> ) -> Option<&T::Artifact>

Finds the artifact with matching path and name

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
let output = project.compile().unwrap();
let contract = output.find("src/Greeter.sol", "Greeter").unwrap();

pub fn find_first(&self, contract_name: impl AsRef<str>) -> Option<&T::Artifact>

Finds the first contract with the given name

pub fn remove( &mut self, path: impl AsRef<str>, contract: impl AsRef<str> ) -> Option<T::Artifact>

Finds the artifact with matching path and name

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
let output = project.compile().unwrap();
let contract = output.find("src/Greeter.sol", "Greeter").unwrap();

pub fn remove_first( &mut self, contract_name: impl AsRef<str> ) -> Option<T::Artifact>

Removes the first contract with the given name from the set

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
let mut output = project.compile().unwrap();
let contract = output.remove_first("Greeter").unwrap();

pub fn remove_contract<'a>( &mut self, info: impl Into<ContractInfoRef<'a>> ) -> Option<T::Artifact>

Removes the contract with matching path and name using the <path>:<contractname> pattern where path is optional.

If the path segment is None, then the first matching Contract is returned, see Self::remove_first

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
use ethers_solc::info::ContractInfo;
let mut output = project.compile().unwrap();
let info = ContractInfo::new("src/Greeter.sol:Greeter");
let contract = output.remove_contract(&info).unwrap();

impl ProjectCompileOutput<ConfigurableArtifacts>

pub fn into_contract_bytecodes( self ) -> impl Iterator<Item = (ArtifactId, CompactContractBytecode)>

A helper functions that extracts the underlying CompactContractBytecode from the crate::ConfigurableContractArtifact

Example

use std::collections::btree_map::BTreeMap;
use ethers_solc::artifacts::contract::CompactContractBytecode;
use ethers_solc::{ArtifactId, Project};
use ethers_solc::contracts::ArtifactContracts;

let project = Project::builder().build().unwrap();
let contracts: ArtifactContracts = project.compile().unwrap().into_contract_bytecodes().collect();