Module ethers_solc::report

Subscribe to events in the compiler pipeline

The reporter is the component of the crate::Project::compile() pipeline which is responsible for reporting on specific steps in the process.

By default, the current reporter is a noop that does nothing.

To use another report implementation, it must be set as the current reporter. There are two methods for doing so: with_scoped and try_init. with_scoped sets the reporter for the duration of a scope, while set_global sets a global default report for the entire process.

Structs

BasicStdoutReporter

A Reporter that emits some general information to stdout

NoReporter

A no-op Reporter that does nothing.

Report

Report trace data to a Reporter.

ScopeGuard

A guard that resets the current scoped reporter to the prior scoped reporter when dropped.

SetGlobalReporterError

Returned if setting the global reporter fails.

SolcCompilerIoReporter

Debug Helper type that can be used to write the crate::Solc CompilerInput and CompilerOutput to disk if configured.

Traits

Reporter

Trait representing the functions required to emit information about various steps in the compiler pipeline.

Functions

format_unresolved_imports

Creates a meaningful message for all unresolved imports

get_default

Executes a closure with a reference to this thread’s current reporter.

init

Install this Reporter as the global default.

set_scoped

Sets the reporter as the scoped reporter for the duration of the lifetime of the returned DefaultGuard

try_init

Install this Reporter as the global default if one is not already set.

with_global

Executes a closure with a reference to the Reporter.

with_scoped

Sets this reporter as the scoped reporter for the duration of a closure.