Struct ethers_solc::sources::VersionedSourceFiles

pub struct VersionedSourceFiles(pub BTreeMap<String, Vec<VersionedSourceFile>>);

(source_file path -> SourceFile + solc version)

Tuple Fields
0: BTreeMap<String, Vec<VersionedSourceFile>>
Implementations

impl VersionedSourceFiles

pub fn slash_paths(&mut self)

Converts all \\ separators in all paths to /

pub fn is_empty(&self) -> bool

pub fn len(&self) -> usize

pub fn files(&self) -> impl Iterator<Item = &String> + '_

Returns an iterator over all files

pub fn into_ids(self) -> impl Iterator<Item = (u32, String)>

Returns an iterator over the source files’ ids and path

use std::collections::BTreeMap;
use ethers_solc::sources::VersionedSourceFiles;
let sources: BTreeMap<u32,String> = files.into_ids().collect();

pub fn into_paths(self) -> impl Iterator<Item = (String, u32)>

Returns an iterator over the source files’ paths and ids

use std::collections::BTreeMap;
use ethers_solc::artifacts::SourceFiles;
let sources :BTreeMap<String, u32> = files.into_paths().collect();

pub fn into_ids_with_version( self ) -> impl Iterator<Item = (u32, String, Version)>

Returns an iterator over the source files’ ids and path

use std::collections::BTreeMap;
use semver::Version;
use ethers_solc::sources::VersionedSourceFiles;
let sources: BTreeMap<(u32, Version) ,String> = files.into_ids_with_version().map(|(id, source, version)|((id, version), source)).collect();

pub fn find_file(&self, source_file: impl AsRef<str>) -> Option<&SourceFile>

Finds the first source file with the given path

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
let output = project.compile().unwrap().output();
let source_file = output.sources.find_file("src/Greeter.sol").unwrap();

pub fn find_file_and_version( &self, path: &str, version: &Version ) -> Option<&SourceFile>

Same as Self::find_file but also checks for version

pub fn find_id(&self, id: u32) -> Option<&SourceFile>

Finds the first source file with the given id

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
let output = project.compile().unwrap().output();
let source_file = output.sources.find_id(0).unwrap();

pub fn find_id_and_version( &self, id: u32, version: &Version ) -> Option<&SourceFile>

Same as Self::find_id but also checks for version

pub fn remove_by_path( &mut self, source_file: impl AsRef<str> ) -> Option<SourceFile>

Removes the first source_file with the given path from the set

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
let (mut sources, _) = project.compile().unwrap().output().split();
let source_file = sources.remove_by_path("src/Greeter.sol").unwrap();

pub fn remove_by_id(&mut self, id: u32) -> Option<SourceFile>

Removes the first source_file with the given id from the set

Example

use ethers_solc::Project;
use ethers_solc::artifacts::*;
let (mut sources, _) = project.compile().unwrap().output().split();
let source_file = sources.remove_by_id(0).unwrap();

pub fn sources(&self) -> impl Iterator<Item = (&String, &SourceFile)>

Iterate over all contracts and their names

pub fn sources_with_version( &self ) -> impl Iterator<Item = (&String, &SourceFile, &Version)>

Returns an iterator over (file, SourceFile, Version)

pub fn into_sources(self) -> impl Iterator<Item = (String, SourceFile)>

Returns an iterator over all contracts and their source names.

use std::collections::BTreeMap;
use ethers_solc::{ artifacts::* };
use ethers_solc::sources::VersionedSourceFiles;
let sources: BTreeMap<String, SourceFile> = sources
    .into_sources()
    .collect();

pub fn into_sources_with_version( self ) -> impl Iterator<Item = (String, SourceFile, Version)>

Returns an iterator over all contracts and their source names.

use std::collections::BTreeMap;
use semver::Version;
use ethers_solc::{ artifacts::* };
use ethers_solc::sources::VersionedSourceFiles;
let sources: BTreeMap<(String,Version), SourceFile> = sources
    .into_sources_with_version().map(|(path, source, version)|((path,version), source))
    .collect();

pub fn join_all(&mut self, root: impl AsRef<Path>) -> &mut Self

Sets the sources’ file paths to root adjoined to self.file.

pub fn strip_prefix_all(&mut self, base: impl AsRef<Path>) -> &mut Self

Removes base from all source file paths