Struct ethers_solc::ProjectBuilder

pub struct ProjectBuilder<T: ArtifactOutput = ConfigurableArtifacts> {
    pub ignored_error_codes: Vec<u64>,
    /* private fields */
}

Fields
ignored_error_codes: Vec<u64>

Which error codes to ignore

Implementations

impl<T: ArtifactOutput> ProjectBuilder<T>

pub fn new(artifacts: T) -> Self

Create a new builder with the given artifacts handler

pub fn paths(self, paths: ProjectPathsConfig) -> Self

pub fn solc(self, solc: impl Into<Solc>) -> Self

pub fn solc_config(self, solc_config: SolcConfig) -> Self

pub fn ignore_error_code(self, code: u64) -> Self

pub fn ignore_error_codes(self, codes: impl IntoIterator<Item = u64>) -> Self

pub fn set_compiler_severity_filter( self, compiler_severity_filter: Severity ) -> Self

pub fn ephemeral(self) -> Self

Disables cached builds

pub fn set_cached(self, cached: bool) -> Self

Sets the cache status

pub fn set_build_info(self, build_info: bool) -> Self

Sets the build info value

pub fn offline(self) -> Self

Activates offline mode

Prevents network possible access to download/check solc installs

pub fn set_offline(self, offline: bool) -> Self

Sets the offline status

pub fn set_slashed_paths(self, slashed_paths: bool) -> Self

Sets whether to slash all paths on windows

If set to true all \\ separators are replaced with /, same as solc

pub fn no_artifacts(self) -> Self

Disables writing artifacts to disk

pub fn set_no_artifacts(self, artifacts: bool) -> Self

Sets the no artifacts status

pub fn set_auto_detect(self, auto_detect: bool) -> Self

Sets automatic solc version detection

pub fn no_auto_detect(self) -> Self

Disables automatic solc version detection

pub fn solc_jobs(self, jobs: usize) -> Self

Sets the maximum number of parallel solc processes to run simultaneously.

Panics

jobs must be at least 1

pub fn single_solc_jobs(self) -> Self

Sets the number of parallel solc processes to 1, no parallelization

pub fn artifacts<A: ArtifactOutput>(self, artifacts: A) -> ProjectBuilder<A>

Set arbitrary ArtifactOutputHandler

pub fn allowed_path<P: Into<PathBuf>>(self, path: P) -> Self

Adds an allowed-path to the solc executable

pub fn allowed_paths<I, S>(self, args: I) -> Self
where I: IntoIterator<Item = S>, S: Into<PathBuf>,

Adds multiple allowed-path to the solc executable

pub fn include_path<P: Into<PathBuf>>(self, path: P) -> Self

Adds an --include-path to the solc executable

pub fn include_paths<I, S>(self, args: I) -> Self
where I: IntoIterator<Item = S>, S: Into<PathBuf>,

Adds multiple include-path to the solc executable

pub fn build(self) -> Result<Project<T>>