Trait ethers_solc::Artifact

pub trait Artifact {

    // Required methods
    fn into_inner(self) -> (Option<Abi>, Option<Bytes>);
    fn into_compact_contract(self) -> CompactContract;
    fn into_contract_bytecode(self) -> CompactContractBytecode;
    fn into_parts(self) -> (Option<Abi>, Option<Bytes>, Option<Bytes>);
    fn get_contract_bytecode(&self) -> CompactContractBytecodeCow<'_>;

    // Provided methods
    fn into_abi(self) -> Option<Abi>
       where Self: Sized { ... }
    fn into_bytecode_bytes(self) -> Option<Bytes>
       where Self: Sized { ... }
    fn into_deployed_bytecode_bytes(self) -> Option<Bytes>
       where Self: Sized { ... }
    fn try_into_parts(self) -> Result<(Abi, Bytes, Bytes)>
       where Self: Sized { ... }
    fn get_bytecode(&self) -> Option<Cow<'_, CompactBytecode>> { ... }
    fn get_bytecode_object(&self) -> Option<Cow<'_, BytecodeObject>> { ... }
    fn get_bytecode_bytes(&self) -> Option<Cow<'_, Bytes>> { ... }
    fn get_deployed_bytecode(&self) -> Option<Cow<'_, CompactDeployedBytecode>> { ... }
    fn get_deployed_bytecode_object(&self) -> Option<Cow<'_, BytecodeObject>> { ... }
    fn get_deployed_bytecode_bytes(&self) -> Option<Cow<'_, Bytes>> { ... }
    fn get_abi(&self) -> Option<Cow<'_, Abi>> { ... }
    fn get_source_map(&self) -> Option<Result<SourceMap, SyntaxError>> { ... }
    fn get_source_map_str(&self) -> Option<Cow<'_, str>> { ... }
    fn get_source_map_deployed(&self) -> Option<Result<SourceMap, SyntaxError>> { ... }
    fn get_source_map_deployed_str(&self) -> Option<Cow<'_, str>> { ... }

}

A trait representation for a crate::Contract artifact

Required Methods

fn into_inner(self) -> (Option<Abi>, Option<Bytes>)

Returns the artifact’s Abi and bytecode

fn into_compact_contract(self) -> CompactContract

Turns the artifact into a container type for abi, compact bytecode and deployed bytecode

fn into_contract_bytecode(self) -> CompactContractBytecode

Turns the artifact into a container type for abi, full bytecode and deployed bytecode

fn into_parts(self) -> (Option<Abi>, Option<Bytes>, Option<Bytes>)

Returns the contents of this type as a single tuple of abi, bytecode and deployed bytecode

fn get_contract_bytecode(&self) -> CompactContractBytecodeCow<'_>

Returns the reference of container type for abi, compact bytecode and deployed bytecode if available

Provided Methods

fn into_abi(self) -> Option<Abi>
where Self: Sized,

Consumes the type and returns the Abi

fn into_bytecode_bytes(self) -> Option<Bytes>
where Self: Sized,

Consumes the type and returns the bytecode

fn into_deployed_bytecode_bytes(self) -> Option<Bytes>
where Self: Sized,

Consumes the type and returns the deployed bytecode

fn try_into_parts(self) -> Result<(Abi, Bytes, Bytes)>
where Self: Sized,

Same as Self::into_parts() but returns Err if an element is None

fn get_bytecode(&self) -> Option<Cow<'_, CompactBytecode>>

Returns the reference to the bytecode

fn get_bytecode_object(&self) -> Option<Cow<'_, BytecodeObject>>

Returns the reference to the bytecode object

fn get_bytecode_bytes(&self) -> Option<Cow<'_, Bytes>>

Returns the bytes of the bytecode object

fn get_deployed_bytecode(&self) -> Option<Cow<'_, CompactDeployedBytecode>>

Returns the reference to the deployedBytecode

fn get_deployed_bytecode_object(&self) -> Option<Cow<'_, BytecodeObject>>

Returns the reference to the bytecode object

fn get_deployed_bytecode_bytes(&self) -> Option<Cow<'_, Bytes>>

Returns the bytes of the deployed bytecode object

fn get_abi(&self) -> Option<Cow<'_, Abi>>

Returns the reference to the Abi if available

fn get_source_map(&self) -> Option<Result<SourceMap, SyntaxError>>

Returns the sourceMap of the creation bytecode

Returns None if no sourceMap string was included in the compiler output Returns Some(Err) if parsing the sourcemap failed

fn get_source_map_str(&self) -> Option<Cow<'_, str>>

Returns the creation bytecode sourceMap as str if it was included in the compiler output

fn get_source_map_deployed(&self) -> Option<Result<SourceMap, SyntaxError>>

Returns the sourceMap of the runtime bytecode

Returns None if no sourceMap string was included in the compiler output Returns Some(Err) if parsing the sourcemap failed

fn get_source_map_deployed_str(&self) -> Option<Cow<'_, str>>

Returns the runtime bytecode sourceMap as str if it was included in the compiler output